Finish draft for flow- and congestion control
I suggest we require the server to finish each congestion window before
starting a new one, because I think it should be fine, if a server
finishes one window before starting a new one, even if a client has
requested more data in between. This solves the problem for now, but I'm
not sure If I'm correct. If I'm wrong, we can of course still change
this.